Refactor according to Dart app best practices
