+++ REFACTORING.md
@@ -1,0 +1,126 @@
+# Refactoring Summary
+
+## Overview
+The codebase has been refactored to improve maintainability, testability, and scalability by separating concerns into distinct modules.
+
+## New Directory Structure
+
+```
+lib/
+├── main.dart              (App entry point - minimal, clean)
+├── config/
+│   └── app_config.dart    (App-wide constants & configuration)
+├── pages/
+│   └── home_page.dart     (Main page layout & keyboard handling)
+├── services/
+│   ├── button_service.dart    (Button action logic)
+│   └── keyboard_service.dart  (Keyboard input handling)
+└── widgets/               (Existing UI components)
+    ├── lcd_display.dart
+    ├── egg_shape.dart
+    └── button.dart
+```
+
+## What Changed
+
+### 1. **main.dart** (from 247 lines → 22 lines)
+**Before:** Mixed app setup, page logic, keyboard handling, button actions
+**After:** Only contains `MyApp` class that sets up the MaterialApp
+
+- Imports `AppConfig` for theme and configuration
+- Imports `HomePage` from pages
+- Uses `AppConfig.getAppTheme()` for cleaner theme setup
+- Much easier to read and maintain
+
+### 2. **lib/config/app_config.dart** (NEW)
+Centralized configuration for the entire app:
+- Color constants (`darkGray`, `primaryRed`)
+- Asset paths (`eggAssetPath`)
+- App information (`appTitle`, `appDescription`)
+- Theme configuration via `getAppTheme()`
+
+**Benefits:**
+- Easy to update colors, strings, and settings app-wide
+- Single source of truth for configuration
+- Reduces magic strings throughout the codebase
+
+### 3. **lib/pages/home_page.dart** (NEW)
+Extracted from `main.dart`, now contains:
+- `HomePage` StatefulWidget with full page layout
+- `_HomePageState` with keyboard and button handling
+- All UI composition logic
+- Injection of services
+
+**Benefits:**
+- Separates page logic from app setup
+- Easier to add new pages in the future
+- Can test page logic independently
+
+### 4. **lib/services/button_service.dart** (NEW)
+Dedicated service for button action logic:
+- `triggerButtonA()`, `triggerButtonB()`, `triggerButtonC()` methods
+- Takes `Logger` as dependency
+- Single responsibility: manage button actions
+
+**Benefits:**
+- Business logic separated from UI
+- Easily testable (inject mock logger)
+- Can be reused in different contexts
+- Easy to add new button logic
+
+### 5. **lib/services/keyboard_service.dart** (NEW)
+Dedicated service for keyboard input handling:
+- `handleKeyPress(KeyEvent)` method
+- Maps keyboard keys (A, B, C) to button actions
+- Depends on `ButtonService` for action execution
+
+**Benefits:**
+- Keyboard logic isolated and testable
+- Can extend to handle more key mappings
+- Decoupled from UI layer
+- Easy to add new keyboard shortcuts
+
+## Refactoring Benefits
+
+✅ **Separation of Concerns**
+- Each file has a single, clear responsibility
+- Easier to understand code flow
+
+✅ **Maintainability**
+- Changes to configuration only in `app_config.dart`
+- Button logic only in `button_service.dart`
+- Page layout only in `home_page.dart`
+
+✅ **Testability**
+- Services can be unit tested independently
+- Can inject mocks/stubs for testing
+- No need to test entire app for small changes
+
+✅ **Scalability**
+- Easy to add new pages (create new file in `pages/`)
+- Easy to add new services
+- Easy to add new button actions
+
+✅ **Reusability**
+- `ButtonService` and `KeyboardService` can be used in other pages
+- Configuration in `AppConfig` available everywhere
+
+## Next Steps for Further Enhancement
+
+1. **State Management**: Add Provider/Riverpod for shared state
+2. **Dependency Injection**: Use GetIt for service management
+3. **Repository Pattern**: Add data access layer if needed
+4. **Testing**: Add unit tests for services
+5. **Localization**: Create a locale configuration file
+6. **Feature Modules**: Organize code by features as app grows
+
+## How to Use This Structure
+
+When adding new features:
+1. Add new buttons? → Update `ButtonService`
+2. Add new keyboard shortcuts? → Update `KeyboardService`
+3. Add new colors/constants? → Update `AppConfig`
+4. Add new pages? → Create new file in `pages/`
+5. Add business logic? → Create new service in `services/`
+
+This structure scales well as your app grows!
